fastapi-demo/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
opentelemetry-api==1.28.0
opentelemetry-sdk==1.28.0
opentelemetry-exporter-otlp==1.28.0
opentelemetry-instrumentation-fastapi==0.49b0
opentelemetry-instrumentation-requests==0.49b0